DE/requirements.txt
absl-py==1.2.0
alembic==1.8.1
appdirs==1.4.4
asciitree==0.3.3
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
audioread==3.0.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bokeh==2.4.3
bordr==0.1.4
botok==0.8.10
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
colorcet==3.0.0
commonmark==0.9.1
cycler==0.11.0
das==0.26.9
dash==2.6.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2022.9.0
datashader==0.14.2
datashape==0.5.2
decorator==5.1.1
defopt==6.4.0
distributed==2022.9.0
Django==4.1.1
dnspython==2.2.1
docutils==0.19
email-validator==1.2.1
entrypoints==0.4
executing==1.0.0
fasteners==0.17.3
flammkuchen==1.0.2
Flask==2.2.2
Flask-Compress==1.12
Flask-Migrate==3.1.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==2.0.7
fonttools==4.37.1
fsspec==2022.8.2
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.48.1
gunicorn==20.1.0
h5py==3.7.0
HeapDict==1.0.1
idna==3.3
image==1.5.33
imageio==2.21.2
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kiwisolver==1.4.4
kt-legacy==1.0.4
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.1
locket==1.0.0
Mako==1.2.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
matplotlib-scalebar==0.8.1
msgpack==1.0.4
multipledispatch==0.6.0
mysqlclient==2.1.1
networkx==2.8.6
numba==0.56.2
numcodecs==0.10.2
numexpr==2.8.3
numpy==1.23.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
param==1.12.2
parso==0.8.3
partd==1.3.0
PeakUtils==1.3.4
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
plotly==5.10.0
pluggy==1.0.0
pockets==0.9.1
pooch==1.6.0
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil==5.9.2
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybo==0.8.0
pycparser==2.21
pyct==0.4.8
pyewts==0.1.3
Pygments==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.0
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
sphinxcontrib-napoleon==0.7
SQLAlchemy==1.4.40
sqlparse==0.4.2
stack-data==0.5.0
tables==3.7.0
tblib==1.7.0
tenacity==8.0.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tibetan-sort==0.2.0
tifffile==2022.8.12
tomli==2.0.1
toolz==0.12.0
tornado==6.1
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
WTForms==3.0.1
xarray==2022.6.0
zarr==2.12.0
zict==2.2.0
zipp==3.8.1